+++ fixesJW.js
@@ -1,0 +1,4 @@
+const JW_TITLE_FIX_DICT = new Map();  //not org title
+JW_TITLE_FIX_DICT.set('Interstelar', '인터스텔라');
+
+window.JW_TITLE_FIX_DICT = JW_TITLE_FIX_DICT;  //export